requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0